requirements.txt
Flask==2.3.4
PyJWT==2.8.0
python-dotenv==1.0.0
flask-cors==4.0.0
